requirements.txt
attrs==21.2.0
Automat==20.2.0
awscli==1.22.34
Babel==2.8.0
bcrypt==3.2.0
black==23.7.0
blinker==1.4
botocore==1.23.34
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==23.1.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.7
distro==1.7.0
distro-info===1.1build1
docutils==0.17.1
ec2-hibinit-agent==1.0.0
filelock==3.12.2
gitdb==4.0.11
GitPython==3.1.41
hibagent==1.0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
mypy-extensions==1.0.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==23.1
parso==0.8.1
pathspec==0.11.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.10.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.21
pyelftools==0.30
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
roman==3.3
rsa==4.8
s3transfer==0.5.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
smmap==5.0.1
sos==4.4
ssh-import-id==5.11
systemd-python==234
tomli==2.0.1
Twisted==22.1.0
typing-extensions==3.10.0.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.24.2
wadllib==1.3.6
zipp==1.0.0
zope.interface==5.4.0
